fm: fix DAC data type
DAC samples are distorded when using Cyclone because most of them are
converted to negative int values.

+++ md.h
@@ -498,7 +498,7 @@ private:
   int fm_ticker[4];
   signed short fm_reg[2][0x100]; // All of them (-1 = not def'd yet)
 
-  int dac_data[0x138], dac_enabled, dac_last;
+  unsigned int dac_data[0x138], dac_enabled, dac_last;
   // Since dac values are normally shifted << 6, 1 is used to mark unchanged
   // values.
   void dac_init() { dac_last = 1; dac_enabled = 1; dac_clear(); }
@@ -507,10 +507,10 @@ private:
     for(int i = 1; i < 0x138; ++i)
       dac_data[i] = 1;
   }
-  void dac_submit(int d) {
+  void dac_submit(uint8_t d) {
     dac_last = (d - 0x80) << 6; if(dac_enabled) dac_data[ras] = dac_last;
   }
-  void dac_enable(int d) {
+  void dac_enable(uint8_t d) {
     dac_enabled = d & 0x80;
     dac_data[ras] = (dac_enabled? dac_last : 1);
   }
